Had to include my enquiry page in the project
i had to fix so many mistakes firstly i had to fix my css there were a lot to mistakes in there and it wasn't corresponding with my html and java.sc. i also fixed the form layout on the enquiry form and i am currently trying the same with my contact form it took me the whole day to do this. The enquiry page didn't take that long to do but fixing the css and java.sc took me almost the whole day validating it and making those they all work well together