Fixed the layout issues in contact page
Updated my Readme File and I also position my code well and made the layout look better and improved. In my feedback, I saw that my contact page positioning was not well well so i fixed it. My contact page wasn't working at all today because of some minor formatting and structure issues in my in my java.sc. The other mistake i made was that I used the same ID as the enquiry form for the contact page so i had to make a page for both of them on my java.sc. And I used the same code i used for the enquiry page for the contact page in my CSS.